refactor: Refactor files
Refs:SCRUM-13
