+++ src/main/java/com/svitsmachnogo/api/SvitsmachnogoApplication.java
@@ -1,4 +1,4 @@
-package com.svitsmachnogo.api.svitsmachnogo;
+package com.svitsmachnogo.api;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
